Rework batch messages so they are added to the registry.
Why? If they are not in the registry the message batch handler will not
be able to properly parse these handlers as it will fall back to use the
URIPubSubMessage Codec and ommit fields.
