+++ microcosm_pubsub/conventions.py
@@ -7,7 +7,7 @@ from inspect import isclass
 from six import string_types
 
 from inflection import underscore
-from marshmallow import fields, Schema
+from marshmallow import fields
 
 from microcosm_pubsub.codecs import PubSubMessageSchema
 
@@ -147,28 +147,3 @@ def stopped(resource, **kwargs):
 
     """
     return make_media_type(resource, lifecycle_change=LifecycleChange.Stopped, **kwargs)
-
-
-class BatchedMessageSchema(Schema):
-    """
-    A wrapper for a single message that is to be published in a
-    MessageBatchSchema.
-
-    """
-    media_type = fields.String(required=True)
-    message = fields.Raw(required=True)
-    topic_arn = fields.String(required=True)
-    opaque_data = fields.Raw(required=True)
-
-
-class MessageBatchSchema(PubSubMessageSchema):
-    """
-    A message indicating that a batch of messages needs to be published.
-
-    """
-    MEDIA_TYPE = created("batch_message")
-
-    messages = fields.List(fields.Nested(BatchedMessageSchema), required=True)
-
-    def deserialize_media_type(self, obj):
-        return MessageBatchSchema.MEDIA_TYPE
